+++ ydb/core/security/secret/ya.make
@@ -1,0 +1,10 @@
+LIBRARY()
+
+PEERDIR(
+    ydb/core/base
+    ydb/library/actors/core
+    yql/essentials/public/issue
+    ydb/public/api/protos
+)
+
+END()
